refactor commandexecutors into local and remote
