+++ backend/src/command_runner.rs
@@ -1,20 +1,51 @@
-use std::{
-    pin::Pin,
-    process::Stdio,
-    task::{Context, Poll},
-    time::Duration,
-};
-
-use command_group::{AsyncCommandGroup, AsyncGroupChild};
-#[cfg(unix)]
-use nix::{
-    sys::signal::{killpg, Signal},
-    unistd::{getpgid, Pid},
-};
+use async_trait::async_trait;
 use serde::{Deserialize, Serialize};
-use tokio::{io::AsyncRead, process::Command};
+use tokio::io::AsyncRead;
 
 use crate::models::Environment;
+
+mod local;
+mod remote;
+
+pub use local::LocalCommandExecutor;
+pub use remote::RemoteCommandExecutor;
+
+// Core trait that defines the interface for command execution
+#[async_trait]
+pub trait CommandExecutor: Send + Sync {
+    /// Start a process and return a handle to it
+    async fn start(
+        &self,
+        request: &CreateCommandRequest,
+    ) -> Result<Box<dyn ProcessHandle>, CommandError>;
+
+    /// Get the executor type for debugging/logging
+    fn executor_type(&self) -> &'static str;
+}
+
+// Trait for managing running processes
+#[async_trait]
+pub trait ProcessHandle: Send + Sync {
+    /// Check if the process is still running, return exit status if finished
+    async fn try_wait(&mut self) -> Result<Option<CommandExitStatus>, CommandError>;
+
+    /// Wait for the process to complete and return exit status
+    async fn wait(&mut self) -> Result<CommandExitStatus, CommandError>;
+
+    /// Kill the process
+    async fn kill(&mut self) -> Result<(), CommandError>;
+
+    /// Get streams for stdout and stderr
+    fn stream(&mut self) -> Result<CommandStream, CommandError>;
+
+    /// Get process identifier (for debugging/logging)
+    fn process_id(&self) -> String;
+
+    /// Check current status (alias for try_wait for backward compatibility)
+    async fn status(&mut self) -> Result<Option<CommandExitStatus>, CommandError> {
+        self.try_wait().await
+    }
+}
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct CreateCommandRequest {
@@ -25,15 +56,8 @@ pub struct CreateCommandRequest {
     pub stdin: Option<String>,
 }
 
-#[derive(Debug, Clone)]
-pub enum CommandRunnerType {
-    Local,
-    Remote,
-}
-
-#[derive(Debug, Clone)]
 pub struct CommandRunner {
-    runner_type: CommandRunnerType,
+    executor: Box<dyn CommandExecutor>,
     command: Option<String>,
     args: Vec<String>,
     working_dir: Option<String>,
@@ -41,18 +65,16 @@ pub struct CommandRunner {
     stdin: Option<String>,
 }
 
-#[derive(Debug)]
-pub enum ProcessHandle {
-    Local(AsyncGroupChild),
-    Remote {
-        process_id: String,
-        cloud_server_url: String,
-    },
+pub struct CommandProcess {
+    handle: Box<dyn ProcessHandle>,
 }
 
-#[derive(Debug)]
-pub struct CommandProcess {
-    handle: Option<ProcessHandle>,
+impl std::fmt::Debug for CommandProcess {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("CommandProcess")
+            .field("process_id", &self.handle.process_id())
+            .finish()
+    }
 }
 
 #[derive(Debug)]
@@ -220,111 +242,6 @@ impl CommandStream {
     }
 }
 
-/// HTTP-based AsyncRead wrapper for true streaming
-pub struct HTTPStream {
-    stream: Pin<Box<dyn futures_util::Stream<Item = Result<Vec<u8>, reqwest::Error>> + Send>>,
-    current_chunk: Vec<u8>,
-    chunk_position: usize,
-    finished: bool,
-}
-
-// HTTPStream needs to be Unpin to work with the AsyncRead trait bounds
-impl Unpin for HTTPStream {}
-
-impl HTTPStream {
-    pub async fn new(url: String) -> Result<Self, CommandError> {
-        let client = reqwest::Client::new();
-        let response = client
-            .get(&url)
-            .send()
-            .await
-            .map_err(|e| CommandError::IoError {
-                error: std::io::Error::other(e),
-            })?;
-
-        if !response.status().is_success() {
-            return Err(CommandError::IoError {
-                error: std::io::Error::other(format!(
-                    "HTTP request failed with status: {}",
-                    response.status()
-                )),
-            });
-        }
-
-        // Use chunk() method to create a stream
-        Ok(Self {
-            stream: Box::pin(futures_util::stream::unfold(
-                response,
-                |mut resp| async move {
-                    match resp.chunk().await {
-                        Ok(Some(chunk)) => Some((Ok(chunk.to_vec()), resp)),
-                        Ok(None) => None,
-                        Err(e) => Some((Err(e), resp)),
-                    }
-                },
-            )),
-            current_chunk: Vec::new(),
-            chunk_position: 0,
-            finished: false,
-        })
-    }
-}
-
-impl AsyncRead for HTTPStream {
-    fn poll_read(
-        mut self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-        buf: &mut tokio::io::ReadBuf<'_>,
-    ) -> Poll<std::io::Result<()>> {
-        if self.finished {
-            return Poll::Ready(Ok(()));
-        }
-
-        // First, try to read from current chunk if available
-        if self.chunk_position < self.current_chunk.len() {
-            let remaining_in_chunk = self.current_chunk.len() - self.chunk_position;
-            let to_read = std::cmp::min(remaining_in_chunk, buf.remaining());
-
-            let chunk_data =
-                &self.current_chunk[self.chunk_position..self.chunk_position + to_read];
-            buf.put_slice(chunk_data);
-            self.chunk_position += to_read;
-
-            return Poll::Ready(Ok(()));
-        }
-
-        // Current chunk is exhausted, try to get the next chunk
-        match self.stream.as_mut().poll_next(cx) {
-            Poll::Ready(Some(Ok(chunk))) => {
-                if chunk.is_empty() {
-                    // Empty chunk, mark as finished
-                    self.finished = true;
-                    Poll::Ready(Ok(()))
-                } else {
-                    // New chunk available
-                    self.current_chunk = chunk;
-                    self.chunk_position = 0;
-
-                    // Read from the new chunk
-                    let to_read = std::cmp::min(self.current_chunk.len(), buf.remaining());
-                    let chunk_data = &self.current_chunk[..to_read];
-                    buf.put_slice(chunk_data);
-                    self.chunk_position = to_read;
-
-                    Poll::Ready(Ok(()))
-                }
-            }
-            Poll::Ready(Some(Err(e))) => Poll::Ready(Err(std::io::Error::other(e))),
-            Poll::Ready(None) => {
-                // Stream ended
-                self.finished = true;
-                Poll::Ready(Ok(()))
-            }
-            Poll::Pending => Poll::Pending,
-        }
-    }
-}
-
 impl Default for CommandRunner {
     fn default() -> Self {
         Self::new()
@@ -343,7 +260,7 @@ impl CommandRunner {
 
     pub fn new_local() -> Self {
         Self {
-            runner_type: CommandRunnerType::Local,
+            executor: Box::new(LocalCommandExecutor::new()),
             command: None,
             args: Vec::new(),
             working_dir: None,
@@ -354,7 +271,7 @@ impl CommandRunner {
 
     pub fn new_remote() -> Self {
         Self {
-            runner_type: CommandRunnerType::Remote,
+            executor: Box::new(RemoteCommandExecutor::new()),
             command: None,
             args: Vec::new(),
             working_dir: None,
@@ -437,466 +354,34 @@ impl CommandRunner {
     }
 
     pub async fn start(&self) -> Result<CommandProcess, CommandError> {
-        let command = self.command.as_ref().ok_or(CommandError::NoCommandSet)?;
+        let request = self.to_request().ok_or(CommandError::NoCommandSet)?;
+        let handle = self.executor.start(&request).await?;
 
-        match self.runner_type {
-            CommandRunnerType::Local => {
-                let mut cmd = Command::new(command);
-
-                cmd.args(&self.args)
-                    .kill_on_drop(true)
-                    .stdin(Stdio::piped())
-                    .stdout(Stdio::piped())
-                    .stderr(Stdio::piped());
-
-                if let Some(dir) = &self.working_dir {
-                    cmd.current_dir(dir);
-                }
-
-                for (key, val) in &self.env_vars {
-                    cmd.env(key, val);
-                }
-
-                let mut child = cmd.group_spawn().map_err(|e| CommandError::SpawnFailed {
-                    command: format!("{} {}", command, self.args.join(" ")),
-                    error: e,
-                })?;
-                if let Some(prompt) = &self.stdin {
-                    // Write prompt to stdin safely
-                    if let Some(mut stdin) = child.inner().stdin.take() {
-                        use tokio::io::AsyncWriteExt;
-                        stdin.write_all(prompt.as_bytes()).await?;
-                        stdin.shutdown().await?;
-                    }
-                }
-
-                Ok(CommandProcess {
-                    handle: Some(ProcessHandle::Local(child)),
-                })
-            }
-            CommandRunnerType::Remote => {
-                let cloud_server_url = std::env::var("CLOUD_SERVER_URL")
-                    .unwrap_or_else(|_| "http://localhost:8000".to_string());
-
-                let request = self.to_request().ok_or(CommandError::NoCommandSet)?;
-
-                let client = reqwest::Client::new();
-                let response = client
-                    .post(format!("{}/commands", cloud_server_url))
-                    .json(&request)
-                    .send()
-                    .await
-                    .map_err(|e| CommandError::IoError {
-                        error: std::io::Error::other(e),
-                    })?;
-
-                let result: serde_json::Value =
-                    response.json().await.map_err(|e| CommandError::IoError {
-                        error: std::io::Error::other(e),
-                    })?;
-
-                let process_id =
-                    result["data"]["process_id"]
-                        .as_str()
-                        .ok_or_else(|| CommandError::IoError {
-                            error: std::io::Error::other(format!(
-                                "Missing process_id in response: {}",
-                                result
-                            )),
-                        })?;
-                Ok(CommandProcess {
-                    handle: Some(ProcessHandle::Remote {
-                        process_id: process_id.to_string(),
-                        cloud_server_url,
-                    }),
-                })
-            }
-        }
+        Ok(CommandProcess { handle })
     }
 }
 
 impl CommandProcess {
     #[allow(dead_code)]
     pub async fn status(&mut self) -> Result<Option<CommandExitStatus>, CommandError> {
-        match &mut self.handle {
-            Some(ProcessHandle::Local(child)) => match child
-                .inner()
-                .try_wait()
-                .map_err(|e| CommandError::StatusCheckFailed { error: e })?
-            {
-                Some(status) => Ok(Some(CommandExitStatus::from_local(status))),
-                None => Ok(None),
-            },
-            Some(ProcessHandle::Remote {
-                process_id,
-                cloud_server_url,
-            }) => {
-                // Make HTTP request to get status from cloud server
-                let client = reqwest::Client::new();
-                let response = client
-                    .get(format!(
-                        "{}/commands/{}/status",
-                        cloud_server_url, process_id
-                    ))
-                    .send()
-                    .await
-                    .map_err(|e| CommandError::StatusCheckFailed {
-                        error: std::io::Error::other(e),
-                    })?;
-
-                if !response.status().is_success() {
-                    if response.status() == reqwest::StatusCode::NOT_FOUND {
-                        return Err(CommandError::StatusCheckFailed {
-                            error: std::io::Error::new(
-                                std::io::ErrorKind::NotFound,
-                                "Process not found",
-                            ),
-                        });
-                    } else {
-                        return Err(CommandError::StatusCheckFailed {
-                            error: std::io::Error::other("Status check failed"),
-                        });
-                    }
-                }
-
-                let result: serde_json::Value =
-                    response
-                        .json()
-                        .await
-                        .map_err(|e| CommandError::StatusCheckFailed {
-                            error: std::io::Error::other(e),
-                        })?;
-
-                let data =
-                    result["data"]
-                        .as_object()
-                        .ok_or_else(|| CommandError::StatusCheckFailed {
-                            error: std::io::Error::other("Invalid response format"),
-                        })?;
-
-                let running = data["running"].as_bool().unwrap_or(false);
-
-                if running {
-                    Ok(None) // Still running
-                } else {
-                    // Process completed, extract exit status
-                    let exit_code = data["exit_code"].as_i64().map(|c| c as i32);
-                    let success = data["success"].as_bool().unwrap_or(false);
-
-                    Ok(Some(CommandExitStatus::from_remote(
-                        exit_code,
-                        success,
-                        Some(process_id.clone()),
-                        None,
-                    )))
-                }
-            }
-            None => Err(CommandError::ProcessNotStarted),
-        }
+        self.handle.status().await
     }
 
     pub async fn try_wait(&mut self) -> Result<Option<CommandExitStatus>, CommandError> {
-        match &mut self.handle {
-            Some(ProcessHandle::Local(child)) => match child
-                .inner()
-                .try_wait()
-                .map_err(|e| CommandError::StatusCheckFailed { error: e })?
-            {
-                Some(status) => Ok(Some(CommandExitStatus::from_local(status))),
-                None => Ok(None),
-            },
-            Some(ProcessHandle::Remote {
-                process_id,
-                cloud_server_url,
-            }) => {
-                // try_wait has same behavior as status for remote processes
-                let client = reqwest::Client::new();
-                let response = client
-                    .get(format!(
-                        "{}/commands/{}/status",
-                        cloud_server_url, process_id
-                    ))
-                    .send()
-                    .await
-                    .map_err(|e| CommandError::StatusCheckFailed {
-                        error: std::io::Error::other(e),
-                    })?;
-
-                if !response.status().is_success() {
-                    if response.status() == reqwest::StatusCode::NOT_FOUND {
-                        return Err(CommandError::StatusCheckFailed {
-                            error: std::io::Error::new(
-                                std::io::ErrorKind::NotFound,
-                                "Process not found",
-                            ),
-                        });
-                    } else {
-                        return Err(CommandError::StatusCheckFailed {
-                            error: std::io::Error::other("Status check failed"),
-                        });
-                    }
-                }
-
-                let result: serde_json::Value =
-                    response
-                        .json()
-                        .await
-                        .map_err(|e| CommandError::StatusCheckFailed {
-                            error: std::io::Error::other(e),
-                        })?;
-
-                let data =
-                    result["data"]
-                        .as_object()
-                        .ok_or_else(|| CommandError::StatusCheckFailed {
-                            error: std::io::Error::other("Invalid response format"),
-                        })?;
-
-                let running = data["running"].as_bool().unwrap_or(false);
-
-                if running {
-                    Ok(None) // Still running
-                } else {
-                    // Process completed, extract exit status
-                    let exit_code = data["exit_code"].as_i64().map(|c| c as i32);
-                    let success = data["success"].as_bool().unwrap_or(false);
-
-                    Ok(Some(CommandExitStatus::from_remote(
-                        exit_code,
-                        success,
-                        Some(process_id.clone()),
-                        None,
-                    )))
-                }
-            }
-            None => Err(CommandError::ProcessNotStarted),
-        }
+        self.handle.try_wait().await
     }
 
     pub async fn kill(&mut self) -> Result<(), CommandError> {
-        match &mut self.handle {
-            Some(ProcessHandle::Local(child)) => {
-                // hit the whole process group, not just the leader
-                #[cfg(unix)]
-                {
-                    if let Some(pid) = child.inner().id() {
-                        let pgid = getpgid(Some(Pid::from_raw(pid as i32))).map_err(|e| {
-                            CommandError::KillFailed {
-                                error: std::io::Error::other(e),
-                            }
-                        })?;
-
-                        for sig in [Signal::SIGINT, Signal::SIGTERM, Signal::SIGKILL] {
-                            if let Err(e) = killpg(pgid, sig) {
-                                tracing::warn!(
-                                    "Failed to send signal {:?} to process group {}: {}",
-                                    sig,
-                                    pgid,
-                                    e
-                                );
-                            }
-                            tokio::time::sleep(Duration::from_secs(2)).await;
-                            if child
-                                .inner()
-                                .try_wait()
-                                .map_err(|e| CommandError::StatusCheckFailed { error: e })?
-                                .is_some()
-                            {
-                                break; // gone!
-                            }
-                        }
-                    }
-                }
-
-                // final fallback – command_group already targets the group
-                child
-                    .kill()
-                    .await
-                    .map_err(|e| CommandError::KillFailed { error: e })?;
-                child
-                    .wait()
-                    .await
-                    .map_err(|e| CommandError::KillFailed { error: e })?; // reap
-
-                // Clear the handle after successful kill
-                self.handle = None;
-                Ok(())
-            }
-            Some(ProcessHandle::Remote {
-                process_id,
-                cloud_server_url,
-            }) => {
-                let client = reqwest::Client::new();
-                let response = client
-                    .delete(format!("{}/commands/{}", cloud_server_url, process_id))
-                    .send()
-                    .await
-                    .map_err(|e| CommandError::KillFailed {
-                        error: std::io::Error::other(e),
-                    })?;
-
-                if !response.status().is_success() {
-                    if response.status() == reqwest::StatusCode::NOT_FOUND {
-                        // Process not found, might have already finished - treat as success
-                        self.handle = None;
-                        return Ok(());
-                    }
-
-                    return Err(CommandError::KillFailed {
-                        error: std::io::Error::other(format!(
-                            "Remote kill failed with status: {}",
-                            response.status()
-                        )),
-                    });
-                }
-
-                // Check if server indicates process was already completed
-                if let Ok(result) = response.json::<serde_json::Value>().await {
-                    if let Some(data) = result.get("data") {
-                        if let Some(message) = data.as_str() {
-                            tracing::info!("Kill result: {}", message);
-                        }
-                    }
-                }
-
-                // Clear the handle after successful kill
-                self.handle = None;
-                Ok(())
-            }
-            None => Err(CommandError::ProcessNotStarted),
-        }
+        self.handle.kill().await
     }
 
     pub fn stream(&mut self) -> Result<CommandStream, CommandError> {
-        match &mut self.handle {
-            Some(ProcessHandle::Local(child)) => Ok(CommandStream::from_local(
-                child.inner().stdout.take(),
-                child.inner().stderr.take(),
-            )),
-            Some(ProcessHandle::Remote {
-                process_id,
-                cloud_server_url,
-            }) => {
-                // Create HTTP streams for stdout and stderr using a blocking approach
-                let stdout_url = format!("{}/commands/{}/stdout", cloud_server_url, process_id);
-                let stderr_url = format!("{}/commands/{}/stderr", cloud_server_url, process_id);
-
-                // Use tokio spawn_blocking to avoid the nested runtime issue
-                let stdout = std::thread::spawn({
-                    let url = stdout_url;
-                    move || {
-                        let rt = tokio::runtime::Runtime::new().unwrap();
-                        rt.block_on(HTTPStream::new(url))
-                    }
-                })
-                .join()
-                .map_err(|_| CommandError::IoError {
-                    error: std::io::Error::other("Failed to create stdout stream"),
-                })?;
-
-                let stderr = std::thread::spawn({
-                    let url = stderr_url;
-                    move || {
-                        let rt = tokio::runtime::Runtime::new().unwrap();
-                        rt.block_on(HTTPStream::new(url))
-                    }
-                })
-                .join()
-                .map_err(|_| CommandError::IoError {
-                    error: std::io::Error::other("Failed to create stderr stream"),
-                })?;
-
-                let stdout_stream: Option<Box<dyn AsyncRead + Unpin + Send>> = stdout
-                    .ok()
-                    .map(|s| Box::new(s) as Box<dyn AsyncRead + Unpin + Send>);
-                let stderr_stream: Option<Box<dyn AsyncRead + Unpin + Send>> = stderr
-                    .ok()
-                    .map(|s| Box::new(s) as Box<dyn AsyncRead + Unpin + Send>);
-
-                Ok(CommandStream::from_streams(stdout_stream, stderr_stream))
-            }
-            None => Err(CommandError::ProcessNotStarted),
-        }
+        self.handle.stream()
     }
 
     #[allow(dead_code)]
     pub async fn wait(&mut self) -> Result<CommandExitStatus, CommandError> {
-        match &mut self.handle {
-            Some(ProcessHandle::Local(child)) => {
-                let status = child
-                    .wait()
-                    .await
-                    .map_err(|e| CommandError::KillFailed { error: e })?;
-                Ok(CommandExitStatus::from_local(status))
-            }
-            Some(ProcessHandle::Remote {
-                process_id,
-                cloud_server_url,
-            }) => {
-                // Poll the status endpoint until process completes
-                loop {
-                    let client = reqwest::Client::new();
-                    let response = client
-                        .get(format!(
-                            "{}/commands/{}/status",
-                            cloud_server_url, process_id
-                        ))
-                        .send()
-                        .await
-                        .map_err(|e| CommandError::StatusCheckFailed {
-                            error: std::io::Error::other(e),
-                        })?;
-
-                    if !response.status().is_success() {
-                        if response.status() == reqwest::StatusCode::NOT_FOUND {
-                            return Err(CommandError::StatusCheckFailed {
-                                error: std::io::Error::new(
-                                    std::io::ErrorKind::NotFound,
-                                    "Process not found",
-                                ),
-                            });
-                        } else {
-                            return Err(CommandError::StatusCheckFailed {
-                                error: std::io::Error::other("Status check failed"),
-                            });
-                        }
-                    }
-
-                    let result: serde_json::Value =
-                        response
-                            .json()
-                            .await
-                            .map_err(|e| CommandError::StatusCheckFailed {
-                                error: std::io::Error::other(e),
-                            })?;
-
-                    let data = result["data"].as_object().ok_or_else(|| {
-                        CommandError::StatusCheckFailed {
-                            error: std::io::Error::other("Invalid response format"),
-                        }
-                    })?;
-
-                    let running = data["running"].as_bool().unwrap_or(false);
-
-                    if !running {
-                        // Process completed, extract exit status and return
-                        let exit_code = data["exit_code"].as_i64().map(|c| c as i32);
-                        let success = data["success"].as_bool().unwrap_or(false);
-
-                        return Ok(CommandExitStatus::from_remote(
-                            exit_code,
-                            success,
-                            Some(process_id.clone()),
-                            None,
-                        ));
-                    }
-
-                    // Wait a bit before polling again
-                    tokio::time::sleep(tokio::time::Duration::from_millis(100)).await;
-                }
-            }
-            None => Err(CommandError::ProcessNotStarted),
-        }
+        self.handle.wait().await
     }
 }
